Improve Apple Music API integration: support catalog data, fix volume parsing, and enhance audio fade logic

+++ src/components/composer/song/SongPicker.tsx
@@ -159,7 +159,7 @@ function writeToLocalStorage(volume: number) {
 function readFromLocalStorage(): number | null {
     const item = localStorage.getItem("volume")
     if (item) {
-        const number = parseInt(item)
+        const number = parseFloat(item)
         return isNaN(number) ? null : number
     } else return null
 }
